refractor: Convert memory.js into TypeScript
